requirements.txt
typer[all]
google-genai
pydantic
pandas
openpyxl
pillow
opencv-python
python-dotenv
pyyaml
tqdm